+++ src/lib/st-table/index.ts
@@ -1,2 +1,3 @@
 export { StTableModule } from './st-table.module';
+export { StTableHeader } from './shared/table-header.interface';
 export { Order, ORDER_TYPE } from './shared/order';