+++ src/lib/st-table/shared/table-header.interface.ts
@@ -1,0 +1,4 @@
+export interface StTableHeader {
+   id: string;
+   label: string;
+}